refactor: GithubClientApi 상수 클래스로 관리
