+++ src/main/java/com/integrated/techhub/auth/application/client/GithubClient.java
@@ -1,4 +1,4 @@
-package com.integrated.techhub.auth.application;
+package com.integrated.techhub.auth.application.client;
 
 import com.integrated.techhub.auth.dto.response.OAuthCrewGithubPrResponse;
 import com.integrated.techhub.auth.dto.response.OAuthGithubUsernameResponse;
@@ -6,7 +6,7 @@ import com.integrated.techhub.auth.dto.response.OAuthTokensResponse;
 
 import java.util.List;
 
-public interface OAuthClient {
+public interface GithubClient {
 
     OAuthTokensResponse getGithubTokens(final String code);
     OAuthTokensResponse getNewAccessToken(final String refreshToken);
