+++ src/main/java/com/integrated/techhub/auth/application/client/GithubRestTemplateClient.java
@@ -1,10 +1,12 @@
-package com.integrated.techhub.auth.application;
+package com.integrated.techhub.auth.application.client;
 
 import com.integrated.techhub.auth.dto.request.GithubOAuthTokenRequest;
 import com.integrated.techhub.auth.dto.request.OAuthTokenRefreshRequest;
 import com.integrated.techhub.auth.dto.response.OAuthCrewGithubPrResponse;
 import com.integrated.techhub.auth.dto.response.OAuthGithubUsernameResponse;
 import com.integrated.techhub.auth.dto.response.OAuthTokensResponse;
+import com.integrated.techhub.auth.util.GithubApiConstants.Auth;
+import com.integrated.techhub.auth.util.GithubApiConstants.Member;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpEntity;
@@ -16,17 +18,12 @@ import org.springframework.web.client.RestTemplate;
 import java.util.ArrayList;
 import java.util.List;
 
-import static java.lang.String.*;
 import static java.lang.String.format;
 import static org.springframework.http.HttpMethod.GET;
 
 @Component
-public class GithubOAuthClient implements OAuthClient {
+public class GithubRestTemplateClient implements GithubClient {
 
-    private static final String MEMBER_INFO_URL = "https://api.github.com/user";
-    private static final String GITHUB_TOKEN_URL = "https://github.com/login/oauth/access_token";
-    private static final String GET_NEW_ACCESS_TOKEN = "https://api.github.com/login/oauth/access_token?clientId=%s&client_secret=%s&grant_type=refresh_token&refresh_token=%s";
-    private static final String LIST_PULL_REQUEST_URL = "https://api.github.com/repos/woowacourse/%s/pulls";
     private static final RestTemplate restTemplate = new RestTemplate();
 
     @Value("${spring.security.oauth2.client.registration.github.client-id}")
@@ -38,7 +35,7 @@ public class GithubOAuthClient implements OAuthClient {
     @Override
     public OAuthTokensResponse getGithubTokens(final String code) {
         return restTemplate.postForObject(
-                GITHUB_TOKEN_URL,
+                Auth.GITHUB_TOKEN_URL,
                 new GithubOAuthTokenRequest(clientId, clientSecret, code),
                 OAuthTokensResponse.class
         );
@@ -47,7 +44,7 @@ public class GithubOAuthClient implements OAuthClient {
     @Override
     public OAuthTokensResponse getNewAccessToken(final String refreshToken) {
         return restTemplate.postForObject(
-                GET_NEW_ACCESS_TOKEN,
+                Auth.GET_NEW_ACCESS_TOKEN_URL,
                 new OAuthTokenRefreshRequest(clientId, clientSecret, refreshToken),
                 OAuthTokensResponse.class
         );
@@ -61,7 +58,7 @@ public class GithubOAuthClient implements OAuthClient {
         final HttpEntity<Void> request = new HttpEntity<>(headers);
 
         return restTemplate.exchange(
-                MEMBER_INFO_URL,
+                Member.MEMBER_INFO_URL,
                 GET,
                 request,
                 OAuthGithubUsernameResponse.class
@@ -73,7 +70,7 @@ public class GithubOAuthClient implements OAuthClient {
         final List<OAuthCrewGithubPrResponse> responses = new ArrayList<>();
         int page = 1;
         while (true) {
-            List<OAuthCrewGithubPrResponse> prs = fetchPrs(accessToken, format(LIST_PULL_REQUEST_URL + "?state=all&page=" + page, repo));
+            List<OAuthCrewGithubPrResponse> prs = fetchPrs(accessToken, format(Member.LIST_PULL_REQUEST_URL + "?state=all&page=" + page, repo));
             if (prs.isEmpty()) break;
             responses.addAll(prs);
             page++;
